+++ builder.py
@@ -1,0 +1,3 @@
+import os
+os.system('python setup.py sdist')
+os.system('python setup.py bdist_wheel')