+++ jusho/jusho.py
@@ -3,30 +3,38 @@ import os
 import sys
 
 from typing import List, Tuple, Union
+from dataclasses import dataclass
 
 
+def get_database_path() -> str:
+    if os.path.isfile(os.path.join(os.path.dirname(__file__), 'data/address.db')):
+        return os.path.join(os.path.dirname(__file__), 'data/address.db')
+    elif os.path.isfile(os.path.join(sys.prefix, 'data/address.db')):
+        return os.path.join(sys.prefix, 'data/address.db')
+    elif os.path.isfile(os.path.join(os.path.dirname(__file__).rsplit('lib', 1)[0], 'data/address.db')):
+        return os.path.join(os.path.dirname(__file__).rsplit('lib', 1)[0], 'data/address.db')
+    else:
+        raise FileNotFoundError("database file not found")
+
+
+@dataclass(frozen=True)
 class Address:
-    def __init__(self, admin_division_code: str, old_postal_code: str, postal_code: str, prefecture_kana: str,
-                 city_kana: str, town_area_kana: str,
-                 prefecture_kanji: str, city_kanji: str, town_area_kanji: str, prefecture_eng: str, city_eng: str,
-                 town_area_eng: str,
-                 multiple_postal_code: int, multiple_address: int, has_chome: int, multiple_town_area: int):
-        self.admin_division_code = admin_division_code
-        self.old_postal_code = old_postal_code
-        self.postal_code = postal_code
-        self.prefecture_kana = prefecture_kana
-        self.city_kana = city_kana
-        self.town_area_kana = town_area_kana
-        self.prefecture_kanji = prefecture_kanji
-        self.city_kanji = city_kanji
-        self.town_area_kanji = town_area_kanji
-        self.prefecture_eng = prefecture_eng
-        self.city_eng = city_eng
-        self.town_area_eng = town_area_eng
-        self.multiple_postal_code = bool(multiple_postal_code)
-        self.multiple_address = bool(multiple_address)
-        self.has_chome = bool(has_chome)
-        self.multiple_town_area = bool(multiple_town_area)
+    admin_division_code: str
+    old_postal_code: str
+    postal_code: str
+    prefecture_kana: str
+    city_kana: str
+    town_area_kana: str
+    prefecture_kanji: str
+    city_kanji: str
+    town_area_kanji: str
+    prefecture_eng: str
+    city_eng: str
+    town_area_eng: str
+    multiple_postal_code: bool
+    multiple_address: bool
+    has_chome: bool
+    multiple_town_area: bool
 
     @property
     def to_sql(self):
@@ -50,21 +58,11 @@ class Address:
                                                    self.town_area_kanji, self.prefecture_eng,
                                                    self.city_eng, self.town_area_eng)
 
-    def __repr__(self):
-        return "<Address: 〒{}-{}, {} {} {}({} {} {})>".format(self.postal_code[:3], self.postal_code[3:],
-                                                              self.prefecture_kanji, self.city_kanji,
-                                                              self.town_area_kanji, self.prefecture_eng,
-                                                              self.city_eng, self.town_area_eng)
-
 
 class Jusho:
     def __init__(self):
-        if os.path.isfile(os.path.join(os.path.dirname(__file__), 'data/address.db')):
-            self.conn = sqlite3.connect(os.path.join(os.path.dirname(__file__), 'data/address.db'))
-        elif os.path.exists(os.path.join(sys.prefix, 'data/address.db')):
-            self.conn = sqlite3.connect(os.path.join(sys.prefix, 'data/address.db'))
-        else:
-            raise FileNotFoundError("database file not found")
+        print(get_database_path())
+        self.conn = sqlite3.connect(get_database_path())
         self.c = self.conn.cursor()
 
     def from_postal_code(self, postal_code: str) -> Union[Address, None]:
@@ -78,43 +76,30 @@ class Jusho:
         self.c.execute("SELECT distinct prefecture_kana, prefecture_kanji, prefecture_eng FROM address")
         return self.c.fetchall()
 
-    def cities_from_prefecture(self, prefecture: str, type='kana') -> List[Tuple[str]]:
-        assert type in ('kana', 'kanji', 'eng')
-        if type == 'kana':
-            self.c.execute("SELECT distinct city_kana, city_kanji, city_eng FROM address WHERE prefecture_kana=?",
-                           (prefecture,))
-        elif type == 'kanji':
-            self.c.execute("SELECT distinct city_kana, city_kanji, city_eng FROM address WHERE prefecture_kanji=?",
-                           (prefecture,))
-        else:
-            self.c.execute("SELECT distinct city_kana, city_kanji, city_eng FROM address WHERE prefecture_eng=?",
-                           (prefecture,))
+    def cities_from_prefecture(self, prefecture: str, type_='kana') -> List[Tuple[str]]:
+        if type_ not in ('kana', 'kanji', 'eng'):
+            raise ValueError("type must be one of 'kana', 'kanji', 'eng'")
+
+        self.c.execute(f"SELECT distinct city_kana, city_kanji, city_eng FROM address WHERE prefecture_{type_}=?",
+                       (prefecture,))
         return self.c.fetchall()
 
-    def towns_from_city(self, prefecture: str, city: str, type='kana') -> List[Address]:
-        assert type in ('kana', 'kanji', 'eng')
-        if type == 'kana':
-            self.c.execute("SELECT distinct * FROM address WHERE prefecture_kana=? AND city_kana=?", (prefecture, city))
-        elif type == 'kanji':
-            self.c.execute("SELECT distinct * FROM address WHERE prefecture_kanji=? AND city_kanji=?",
-                           (prefecture, city))
-        else:
-            self.c.execute("SELECT distinct * FROM address WHERE prefecture_eng=? AND city_eng", (prefecture, city))
+    def towns_from_city(self, prefecture: str, city: str, type_='kana') -> List[Address]:
+        if type_ not in ('kana', 'kanji', 'eng'):
+            raise ValueError("type must be one of 'kana', 'kanji', 'eng'")
+
+        self.c.execute(f"SELECT distinct * FROM address WHERE prefecture_{type_}=? AND city_{type_}=?",
+                       (prefecture, city))
         return [Address(*result) for result in self.c.fetchall()]
 
-    def address_from_town(self, prefecture: str, city: str, town: str, type='kana') -> Union[List[Address], List[None]]:
-        assert type in ('kana', 'kanji', 'eng')
-        if type == 'kana':
-            self.c.execute(
-                "SELECT distinct * FROM address WHERE prefecture_kana=? AND city_kana=? AND town_area_kana=?",
-                (prefecture, city, town))
-        elif type == 'kanji':
-            self.c.execute(
-                "SELECT distinct * FROM address WHERE prefecture_kanji=? AND city_kanji=? AND town_area_kanji=?",
-                (prefecture, city, town))
-        else:
-            self.c.execute("SELECT distinct * FROM address WHERE prefecture_eng=? AND city_eng=? AND town_area_eng=?",
-                           (prefecture, city, town))
+    def address_from_town(self, prefecture: str, city: str, town: str, type_='kana') -> Union[
+        List[Address], List[None]]:
+        if type_ not in ('kana', 'kanji', 'eng'):
+            raise ValueError("type must be one of 'kana', 'kanji', 'eng'")
+
+        self.c.execute(
+            f"SELECT distinct * FROM address WHERE prefecture_{type_}=? AND city_{type_}=? AND town_area_{type_}=?",
+            (prefecture, city, town))
         result = self.c.fetchone()
         return Address(*result) if result else [None]